Move bin/glustercli to top-level glustercli package
