+++ glustercli/cmd/peer.go
