+++ glustercli/oldfmt.go
@@ -1,4 +1,4 @@
-package main
+package glustercli
 
 import (
 	"os"
